requirements.txt
flask>=3.0
gunicorn>=22.0
requests>=2.31
weasyprint>=62.0
jinja2>=3.1
anthropic>=0.40.0
python-dotenv>=1.0
